api: Lamp values are now always float.

+++ VisualPinball.Engine.Mpf.Unity/Runtime/MpfGamelogicEngine.cs
@@ -153,7 +153,7 @@ namespace VisualPinball.Engine.Mpf.Unity
 			OnCoilChanged?.Invoke(this, new CoilEventArgs(id, isEnabled));
 		}
 
-		public void SetLamp(string id, int value, bool isCoil = false, LampSource source = LampSource.Lamp)
+		public void SetLamp(string id, float value, bool isCoil = false, LampSource source = LampSource.Lamp)
 		{
 			OnLampChanged?.Invoke(this, new LampEventArgs(id, value, isCoil, source));
 		}
